Uml implementation (#11)
* added_food_value_and_changed_tradable_properties_visibility

* added_maxence_algorithm_wip

* fix_merge_errors

* trader_asking_implemented_bug_in_sorting

* Fix->Tradable/StackExchange:fix_name_vector/fix_insertion_sort

* feat->Trader/StockExchange:TraderSelling/ResolveOffers

+++ EconomicEngine/Food.h
@@ -8,7 +8,7 @@ protected:
 	int foodValue;
 public:
 	Food();
-	Food(int foodValue, int count);
+	Food(int foodValue, float count);
 };
 
 #endif
